Cast json_len to Py_ssize_t before Py_BuildValue("s#", ...)
bsonjs.h defines PY_SSIZE_T_CLEAN before including Python.h, so the
"s#" format code reads a Py_ssize_t out of the varargs list. _dumps()
was passing json_len as a bare size_t, the type bson_str_to_json (and
libbson's bson_as_*_extended_json functions) use for output lengths.
This happened to work on platforms where the two types are the same
width, but was reading the wrong type out of the varargs list.

load()/loads() already use Py_ssize_t for their length variables, so
this was the only mismatched call site.

+++ bsonjs/bsonjs.c
@@ -82,7 +82,7 @@ _dumps(PyObject *bson, int mode)
         return NULL;
     }
 
-    rv = Py_BuildValue("s#", json, json_len);
+    rv = Py_BuildValue("s#", json, (Py_ssize_t)json_len);
     bson_free((void *)json);
     return rv;
 }
